+++ Models/ChartMarkerKind.cs
@@ -1,8 +1,0 @@
-namespace CursorPace.Models;
-
-public enum ChartMarkerKind
-{
-    Origin,
-    Sample,
-    Edit
-}